Reorganized core model interface
