+++ ssm/__init__.py
@@ -1,0 +1,4 @@
+# Default imports for SSM
+
+from .hmm import *
+from .lds import *
