+++ ssm/hmm.py
@@ -1,22 +1,25 @@
-import copy
-import warnings
 from functools import partial
 from tqdm.auto import trange
 
 import autograd.numpy as np
 import autograd.numpy.random as npr
-from autograd.scipy.misc import logsumexp
-from autograd.tracer import getval
-from autograd.misc import flatten
 from autograd import value_and_grad
 
-from ssm.optimizers import adam_step, rmsprop_step, sgd_step, convex_combination
-from ssm.primitives import hmm_normalizer, hmm_expected_states, hmm_filter, hmm_sample, viterbi
-from ssm.util import ensure_args_are_lists, ensure_args_not_none, \
+from .optimizers import adam_step, rmsprop_step, sgd_step, convex_combination
+from .primitives import hmm_normalizer, hmm_expected_states, hmm_filter, hmm_sample, viterbi
+from .util import ensure_args_are_lists, ensure_args_not_none, \
     ensure_slds_args_not_none, ensure_variational_args_are_lists, \
     replicate, collapse
 
-class BaseHMM(object):
+from . import observations as obs
+from . import transitions as trans
+from . import init_state_distns as isd
+from . import hierarchical as hier
+from . import emissions as emssn
+
+__all__ = ['HMM', 'HSMM']
+
+class HMM(object):
     """
     Base class for hidden Markov models.
 
@@ -28,7 +31,91 @@ class BaseHMM(object):
     In the code we will sometimes refer to the discrete
     latent state sequence as z and the data as x.
     """
-    def __init__(self, K, D, M, init_state_distn, transitions, observations):
+    def __init__(self, K, D, *, M=0, init_state_distn=None,
+                 transitions='standard',
+                 transition_kwargs=None,
+                 hierarchical_transition_tags=None,
+                 observations="gaussian", observation_kwargs=None,
+                 hierarchical_observation_tags=None, **kwargs):
+
+        # Make the initial state distribution
+        if init_state_distn is None:
+            init_state_distn = isd.InitialStateDistribution(K, D, M=M)
+        if not isinstance(init_state_distn, isd.InitialStateDistribution):
+            raise TypeError("'init_state_distn' must be a subclass of"
+                            " ssm.init_state_distns.InitialStateDistribution")
+
+        # Make the transition model
+        transition_classes = dict(
+            standard=trans.StationaryTransitions,
+            stationary=trans.StationaryTransitions,
+            sticky=trans.StickyTransitions,
+            inputdriven=trans.InputDrivenTransitions,
+            recurrent=trans.RecurrentTransitions,
+            recurrent_only=trans.RecurrentOnlyTransitions,
+            rbf_recurrent=trans.RBFRecurrentTransitions,
+            nn_recurrent=trans.NeuralNetworkRecurrentTransitions
+            )
+
+        if isinstance(transitions, str):
+            if transitions not in transition_classes:
+                raise Exception("Invalid transition model: {}. Must be one of {}".
+                    format(transitions, list(transition_classes.keys())))
+
+            transition_kwargs = transition_kwargs or {}
+            transitions = \
+                hier.HierarchicalTransitions(transition_classes[transitions], K, D, M=M,
+                                        tags=hierarchical_transition_tags,
+                                        **transition_kwargs) \
+                if hierarchical_transition_tags is not None \
+                else transition_classes[transitions](K, D, M=M, **transition_kwargs)
+        if not isinstance(transitions, trans.Transitions):
+            raise TypeError("'transitions' must be a subclass of"
+                            " ssm.transitions.Transitions")
+
+        # This is the master list of observation classes.
+        # When you create a new observation class, add it here.
+        observation_classes = dict(
+            gaussian=obs.GaussianObservations,
+            diagonal_gaussian=obs.DiagonalGaussianObservations,
+            studentst=obs.MultivariateStudentsTObservations,
+            t=obs.MultivariateStudentsTObservations,
+            diagonal_t=obs.StudentsTObservations,
+            diagonal_studentst=obs.StudentsTObservations,
+            bernoulli=obs.BernoulliObservations,
+            categorical=obs.CategoricalObservations,
+            poisson=obs.PoissonObservations,
+            vonmises=obs.VonMisesObservations,
+            ar=obs.AutoRegressiveObservations,
+            autoregressive=obs.AutoRegressiveObservations,
+            no_input_ar=obs.AutoRegressiveObservationsNoInput,
+            diagonal_ar=obs.AutoRegressiveDiagonalNoiseObservations,
+            diagonal_autoregressive=obs.AutoRegressiveDiagonalNoiseObservations,
+            independent_ar=obs.IndependentAutoRegressiveObservations,
+            robust_ar=obs.RobustAutoRegressiveObservations,
+            no_input_robust_ar=obs.RobustAutoRegressiveObservationsNoInput,
+            robust_autoregressive=obs.RobustAutoRegressiveObservations,
+            diagonal_robust_ar=obs.RobustAutoRegressiveDiagonalNoiseObservations,
+            diagonal_robust_autoregressive=obs.RobustAutoRegressiveDiagonalNoiseObservations,
+            )
+
+        if isinstance(observations, str):
+            observations = observations.lower()
+            if observations not in observation_classes:
+                raise Exception("Invalid observation model: {}. Must be one of {}".
+                    format(observations, list(observation_classes.keys())))
+
+            observation_kwargs = observation_kwargs or {}
+            observations = \
+                hier.HierarchicalObservations(observation_classes[observations], K, D, M=M,
+                                        tags=hierarchical_observation_tags,
+                                        **observation_kwargs) \
+                if hierarchical_observation_tags is not None \
+                else observation_classes[observations](K, D, M=M, **observation_kwargs)
+        if not isinstance(observations, obs.Observations):
+            raise TypeError("'observations' must be a subclass of"
+                            " ssm.observations.Observations")
+        
         self.K, self.D, self.M = K, D, M
         self.init_state_distn = init_state_distn
         self.transitions = transitions
@@ -379,7 +466,7 @@ class BaseHMM(object):
         return _fitting_methods[method](datas, inputs=inputs, masks=masks, tags=tags, **kwargs)
 
 
-class BaseHSMM(BaseHMM):
+class HSMM(HMM):
     """
     Hidden semi-Markov model with non-geometric duration distributions.
     The trick is to expand the state space with "super states" and "sub states"
@@ -388,6 +475,75 @@ class BaseHSMM(BaseHMM):
     super+sub states ((1,1), ..., (1,r_1), ..., (K,1), ..., (K,r_K)).
     Here, r_k denotes the number of sub-states of state k.
     """
+
+    def __init__(self, K, D, *, M=0, init_state_distn=None,
+                 transitions="nb", transition_kwargs=None,
+                 observations="gaussian", observation_kwargs=None,
+                 **kwargs):
+
+        if init_state_distn is None:
+            init_state_distn = isd.InitialStateDistribution(K, D, M=M)
+        if not isinstance(init_state_distn, isd.InitialStateDistribution):
+            raise TypeError("'init_state_distn' must be a subclass of"
+                            " ssm.init_state_distns.InitialStateDistribution")
+
+        # Make the transition model
+        transition_classes = dict(
+            nb=NegativeBinomialSemiMarkovTransitions,
+            )
+        if isinstance(transitions, str):
+            if transitions not in transition_classes:
+                raise Exception("Invalid transition model: {}. Must be one of {}".
+                    format(transitions, list(transition_classes.keys())))
+
+            transition_kwargs = transition_kwargs or {}
+            transitions = transition_classes[transitions](K, D, M=M, **transition_kwargs)
+        if not isinstance(transitions, trans.Transitions):
+            raise TypeError("'transitions' must be a subclass of"
+                            " ssm.transitions.Transitions")
+
+        # This is the master list of observation classes.
+        # When you create a new observation class, add it here.
+        observation_classes = dict(
+            gaussian=obs.GaussianObservations,
+            diagonal_gaussian=obs.DiagonalGaussianObservations,
+            studentst=obs.MultivariateStudentsTObservations,
+            t=obs.MultivariateStudentsTObservations,
+            diagonal_t=obs.StudentsTObservations,
+            diagonal_studentst=obs.StudentsTObservations,
+            bernoulli=obs.BernoulliObservations,
+            categorical=obs.CategoricalObservations,
+            poisson=obs.PoissonObservations,
+            vonmises=obs.VonMisesObservations,
+            ar=obs.AutoRegressiveObservations,
+            autoregressive=obs.AutoRegressiveObservations,
+            diagonal_ar=obs.AutoRegressiveDiagonalNoiseObservations,
+            diagonal_autoregressive=obs.AutoRegressiveDiagonalNoiseObservations,
+            independent_ar=obs.IndependentAutoRegressiveObservations,
+            robust_ar=obs.RobustAutoRegressiveObservations,
+            robust_autoregressive=obs.RobustAutoRegressiveObservations,
+            diagonal_robust_ar=obs.RobustAutoRegressiveDiagonalNoiseObservations,
+            diagonal_robust_autoregressive=obs.RobustAutoRegressiveDiagonalNoiseObservations,
+            )
+
+        if isinstance(observations, str):
+            observations = observations.lower()
+            if observations not in observation_classes:
+                raise Exception("Invalid observation model: {}. Must be one of {}".
+                    format(observations, list(observation_classes.keys())))
+
+            observation_kwargs = observation_kwargs or {}
+            observations = observation_classes[observations](K, D, M=M, **observation_kwargs)
+        if not isinstance(observations, obs.Observations):
+            raise TypeError("'observations' must be a subclass of"
+                            " ssm.observations.Observations")
+
+        super().__init__(K, D, M=M, transitions=transition_distn, 
+                        transition_kwargs=transition_kwargs,
+                        observations=observation_distn,
+                        observation_kwargs=observation_kwargs,
+                        **kwargs)
+    
     @property
     def state_map(self):
         return self.transitions.state_map
@@ -613,508 +769,3 @@ class BaseHSMM(BaseHMM):
             self.initialize(datas, inputs=inputs, masks=masks, tags=tags)
 
         return _fitting_methods[method](datas, inputs=inputs, masks=masks, tags=tags, **kwargs)
-
-
-class BaseSwitchingLDS(object):
-    """
-    Switching linear dynamical system fit with
-    stochastic variational inference on the marginal model,
-    integrating out the discrete states.
-    """
-    def __init__(self, N, K, D, M, init_state_distn, transitions, dynamics, emissions):
-        self.N, self.K, self.D, self.M = N, K, D, M
-        self.init_state_distn = init_state_distn
-        self.transitions = transitions
-        self.dynamics = dynamics
-        self.emissions = emissions
-
-    @property
-    def params(self):
-        return self.init_state_distn.params, \
-               self.transitions.params, \
-               self.dynamics.params, \
-               self.emissions.params
-
-    @params.setter
-    def params(self, value):
-        self.init_state_distn.params = value[0]
-        self.transitions.params = value[1]
-        self.dynamics.params = value[2]
-        self.emissions.params = value[3]
-
-    @ensure_args_are_lists
-    def initialize(self, datas, inputs=None, masks=None, tags=None, num_em_iters=25):
-        # First initialize the observation model
-        self.emissions.initialize(datas, inputs, masks, tags)
-
-        # Get the initialized variational mean for the data
-        xs = [self.emissions.invert(data, input, mask, tag)
-              for data, input, mask, tag in zip(datas, inputs, masks, tags)]
-        xmasks = [np.ones_like(x, dtype=bool) for x in xs]
-
-        # Now run a few iterations of EM on a ARHMM with the variational mean
-        print("Initializing with an ARHMM using {} steps of EM.".format(num_em_iters))
-        arhmm = BaseHMM(self.K, self.D, self.M,
-                     copy.deepcopy(self.init_state_distn),
-                     copy.deepcopy(self.transitions),
-                     copy.deepcopy(self.dynamics))
-
-        arhmm.fit(xs, inputs=inputs, masks=xmasks, tags=tags,
-                  method="em", num_em_iters=num_em_iters)
-
-        self.init_state_distn = copy.deepcopy(arhmm.init_state_distn)
-        self.transitions = copy.deepcopy(arhmm.transitions)
-        self.dynamics = copy.deepcopy(arhmm.observations)
-
-    def permute(self, perm):
-        """
-        Permute the discrete latent states.
-        """
-        assert np.all(np.sort(perm) == np.arange(self.K))
-        self.init_state_distn.permute(perm)
-        self.transitions.permute(perm)
-        self.dynamics.permute(perm)
-        self.emissions.permute(perm)
-
-    def log_prior(self):
-        """
-        Compute the log prior probability of the model parameters
-        """
-        return self.init_state_distn.log_prior() + \
-               self.transitions.log_prior() + \
-               self.dynamics.log_prior() + \
-               self.emissions.log_prior()
-
-    def sample(self, T, input=None, tag=None, prefix=None, with_noise=True):
-        K = self.K
-        D = (self.D,) if isinstance(self.D, int) else self.D
-        M = (self.M,) if isinstance(self.M, int) else self.M
-        assert isinstance(D, tuple)
-        assert isinstance(M, tuple)
-
-        # If prefix is given, pad the output with it
-        if prefix is None:
-            pad = 1
-            z = np.zeros(T+1, dtype=int)
-            x = np.zeros((T+1,) + D)
-            data = np.zeros((T+1,) + D)
-            input = np.zeros((T+1,) + M) if input is None else input
-            xmask = np.ones((T+1,) + D, dtype=bool)
-
-            # Sample the first state from the initial distribution
-            pi0 = np.exp(self.init_state_distn.log_initial_state_distn(data, input, xmask, tag))
-            z[0] = npr.choice(self.K, p=pi0)
-            x[0] = self.dynamics.sample_x(z[0], x[:0], tag=tag, with_noise=with_noise)
-
-        else:
-            zhist, xhist, yhist = prefix
-            pad = len(zhist)
-            assert zhist.dtype == int and zhist.min() >= 0 and zhist.max() < K
-            assert xhist.shape == (pad, D)
-            assert yhist.shape == (pad, N)
-
-            z = np.concatenate((zhist, np.zeros(T, dtype=int)))
-            x = np.concatenate((xhist, np.zeros((T,) + D)))
-            input = np.zeros((T+pad,) + M) if input is None else input
-            xmask = np.ones((T+pad,) + D, dtype=bool)
-
-        # Sample z and x
-        for t in range(pad, T+pad):
-            Pt = np.exp(self.transitions.log_transition_matrices(x[t-1:t+1], input[t-1:t+1], mask=xmask[t-1:t+1], tag=tag))[0]
-            z[t] = npr.choice(self.K, p=Pt[z[t-1]])
-            x[t] = self.dynamics.sample_x(z[t], x[:t], input=input[t], tag=tag, with_noise=with_noise)
-
-        # Sample observations given latent states
-        # TODO: sample in the loop above?
-        y = self.emissions.sample(z, x, input=input, tag=tag)
-        return z[pad:], x[pad:], y[pad:]
-
-    @ensure_slds_args_not_none
-    def expected_states(self, variational_mean, data, input=None, mask=None, tag=None):
-        x_mask = np.ones_like(variational_mean, dtype=bool)
-        log_pi0 = self.init_state_distn.log_initial_state_distn(variational_mean, input, x_mask, tag)
-        log_Ps = self.transitions.log_transition_matrices(variational_mean, input, x_mask, tag)
-        log_likes = self.dynamics.log_likelihoods(variational_mean, input, x_mask, tag)
-        log_likes += self.emissions.log_likelihoods(data, input, mask, tag, variational_mean)
-        return hmm_expected_states(log_pi0, log_Ps, log_likes)
-
-    @ensure_slds_args_not_none
-    def most_likely_states(self, variational_mean, data, input=None, mask=None, tag=None):
-        log_pi0 = self.init_state_distn.log_initial_state_distn(variational_mean, input, mask, tag)
-        log_Ps = self.transitions.log_transition_matrices(variational_mean, input, mask, tag)
-        log_likes = self.dynamics.log_likelihoods(variational_mean, input, np.ones_like(variational_mean, dtype=bool), tag)
-        log_likes += self.emissions.log_likelihoods(data, input, mask, tag, variational_mean)
-        return viterbi(log_pi0, log_Ps, log_likes)
-
-    @ensure_slds_args_not_none
-    def smooth(self, variational_mean, data, input=None, mask=None, tag=None):
-        """
-        Compute the mean observation under the posterior distribution
-        of latent discrete states.
-        """
-        Ez, _, _ = self.expected_states(variational_mean, data, input, mask, tag)
-        return self.emissions.smooth(Ez, variational_mean, data, input, tag)
-
-    @ensure_args_are_lists
-    def log_probability(self, datas, inputs=None, masks=None, tags=None):
-        warnings.warn("Cannot compute exact marginal log probability for the SLDS. "
-                      "the ELBO instead.")
-        return np.nan
-
-    @ensure_variational_args_are_lists
-    def elbo(self, variational_posterior, datas, inputs=None, masks=None, tags=None, n_samples=1):
-        """
-        Lower bound on the marginal likelihood p(y | theta)
-        using variational posterior q(x; phi) where phi = variational_params
-        """
-        elbo = 0
-        for sample in range(n_samples):
-            # Sample x from the variational posterior
-            xs = variational_posterior.sample()
-
-            # log p(theta)
-            elbo += self.log_prior()
-
-            # log p(x, y | theta) = log \sum_z p(x, y, z | theta)
-            for x, data, input, mask, tag in zip(xs, datas, inputs, masks, tags):
-
-                # The "mask" for x is all ones
-                x_mask = np.ones_like(x, dtype=bool)
-                log_pi0 = self.init_state_distn.log_initial_state_distn(x, input, x_mask, tag)
-                log_Ps = self.transitions.log_transition_matrices(x, input, x_mask, tag)
-                log_likes = self.dynamics.log_likelihoods(x, input, x_mask, tag)
-                log_likes += self.emissions.log_likelihoods(data, input, mask, tag, x)
-                elbo += hmm_normalizer(log_pi0, log_Ps, log_likes)
-
-            # -log q(x)
-            elbo -= variational_posterior.log_density(xs)
-            assert np.isfinite(elbo)
-
-        return elbo / n_samples
-
-    @ensure_variational_args_are_lists
-    def _surrogate_elbo(self, variational_posterior, datas, inputs=None, masks=None, tags=None,
-        alpha=0.75, **kwargs):
-        """
-        Lower bound on the marginal likelihood p(y | gamma)
-        using variational posterior q(x; phi) where phi = variational_params
-        and gamma = emission parameters.  As part of computing this objective,
-        we optimize q(z | x) and take a natural gradient step wrt theta, the
-        parameters of the dynamics model.
-
-        Note that the surrogate ELBO is a lower bound on the ELBO above.
-           E_p(z | x, y)[log p(z, x, y)]
-           = E_p(z | x, y)[log p(z, x, y) - log p(z | x, y) + log p(z | x, y)]
-           = E_p(z | x, y)[log p(x, y) + log p(z | x, y)]
-           = log p(x, y) + E_p(z | x, y)[log p(z | x, y)]
-           = log p(x, y) -H[p(z | x, y)]
-          <= log p(x, y)
-        with equality only when p(z | x, y) is atomic.  The gap equals the
-        entropy of the posterior on z.
-        """
-        # log p(theta)
-        elbo = self.log_prior()
-
-        # Sample x from the variational posterior
-        xs = variational_posterior.sample()
-
-        # Inner optimization: find the true posterior p(z | x, y; theta).
-        # Then maximize the inner ELBO wrt theta,
-        #
-        #    E_p(z | x, y; theta_fixed)[log p(z, x, y; theta).
-        #
-        # This can be seen as a natural gradient step in theta
-        # space.  Note: we do not want to compute gradients wrt x or the
-        # emissions parameters backward throgh this optimization step,
-        # so we unbox them first.
-        xs_unboxed = [getval(x) for x in xs]
-        emission_params_boxed = self.emissions.params
-        flat_emission_params_boxed, unflatten = flatten(emission_params_boxed)
-        self.emissions.params = unflatten(getval(flat_emission_params_boxed))
-
-        # E step: compute the true posterior p(z | x, y, theta_fixed) and
-        # the necessary expectations under this posterior.
-        expectations = [self.expected_states(x, data, input, mask, tag)
-                        for x, data, input, mask, tag
-                        in zip(xs_unboxed, datas, inputs, masks, tags)]
-
-        # M step: maximize expected log joint wrt parameters
-        # Note: Only do a partial update toward the M step for this sample of xs
-        x_masks = [np.ones_like(x, dtype=bool) for x in xs_unboxed]
-        for distn in [self.init_state_distn, self.transitions, self.dynamics]:
-            curr_prms = copy.deepcopy(distn.params)
-            distn.m_step(expectations, xs_unboxed, inputs, x_masks, tags, **kwargs)
-            distn.params = convex_combination(curr_prms, distn.params, alpha)
-
-        # Box up the emission parameters again before computing the ELBO
-        self.emissions.params = emission_params_boxed
-
-        # Compute expected log likelihood E_q(z | x, y) [log p(z, x, y; theta)]
-        for (Ez, Ezzp1, _), x, x_mask, data, mask, input, tag in \
-            zip(expectations, xs, x_masks, datas, masks, inputs, tags):
-
-            # Compute expected log likelihood (inner ELBO)
-            log_pi0 = self.init_state_distn.log_initial_state_distn(x, input, x_mask, tag)
-            log_Ps = self.transitions.log_transition_matrices(x, input, x_mask, tag)
-            log_likes = self.dynamics.log_likelihoods(x, input, x_mask, tag)
-            log_likes += self.emissions.log_likelihoods(data, input, mask, tag, x)
-
-            elbo += np.sum(Ez[0] * log_pi0)
-            elbo += np.sum(Ezzp1 * log_Ps)
-            elbo += np.sum(Ez * log_likes)
-
-        # -log q(x)
-        elbo -= variational_posterior.log_density(xs)
-        assert np.isfinite(elbo)
-
-        return elbo
-
-    def _fit_svi(self, variational_posterior, datas, inputs, masks, tags,
-                 learning=True, optimizer="adam", num_iters=100, **kwargs):
-        """
-        Fit with stochastic variational inference using a
-        mean field Gaussian approximation for the latent states x_{1:T}.
-        """
-        # Define the objective (negative ELBO)
-        T = sum([data.shape[0] for data in datas])
-        def _objective(params, itr):
-            if learning:
-                self.params, variational_posterior.params = params
-            else:
-                variational_posterior.params = params
-
-            obj = self.elbo(variational_posterior, datas, inputs, masks, tags)
-            return -obj / T
-
-        # Initialize the parameters
-        if learning:
-            params = (self.params, variational_posterior.params)
-        else:
-            params = variational_posterior.params
-
-        # Set up the progress bar
-        elbos = [-_objective(params, 0) * T]
-        pbar = trange(num_iters)
-        pbar.set_description("ELBO: {:.1f}".format(elbos[0]))
-
-        # Run the optimizer
-        step = dict(sgd=sgd_step, rmsprop=rmsprop_step, adam=adam_step)[optimizer]
-        state = None
-        for itr in pbar:
-            params, val, g, state = step(value_and_grad(_objective), params, itr, state)
-            elbos.append(-val * T)
-
-            # TODO: Check for convergence -- early stopping
-
-            # Update progress bar
-            pbar.set_description("ELBO: {:.1f}".format(elbos[-1]))
-            pbar.update()
-
-        # Save the final parameters
-        if learning:
-            self.params, variational_posterior.params = params
-        else:
-            variational_posterior.params = params
-
-        return elbos
-
-    def _fit_variational_em(self, variational_posterior, datas, inputs, masks, tags,
-                 learning=True, alpha=.75, optimizer="adam", num_iters=100, **kwargs):
-        """
-        Let gamma denote the emission parameters and theta denote the transition
-        and initial discrete state parameters. This is a mix of EM and SVI:
-            1. Sample x ~ q(x; phi)
-            2. Compute L(x, theta') = E_p(z | x, theta)[log p(x, z; theta')]
-            3. Set theta = (1 - alpha) theta + alpha * argmax L(x, theta')
-            4. Set gamma = gamma + eps * nabla log p(y | x; gamma)
-            5. Set phi = phi + eps * dx/dphi * d/dx [L(x, theta) + log p(y | x; gamma) - log q(x; phi)]
-        """
-        # Optimize the standard ELBO when updating gamma (emissions params)
-        # and phi (variational params)
-        T = sum([data.shape[0] for data in datas])
-        def _objective(params, itr):
-            if learning:
-                self.emissions.params, variational_posterior.params = params
-            else:
-                variational_posterior.params = params
-
-            obj = self._surrogate_elbo(variational_posterior, datas, inputs, masks, tags, **kwargs)
-            return -obj / T
-
-        # Initialize the parameters
-        if learning:
-            params = (self.emissions.params, variational_posterior.params)
-        else:
-            params = variational_posterior.params
-
-        # Set up the progress bar
-        elbos = [-_objective(params, 0) * T]
-        pbar = trange(num_iters)
-        pbar.set_description("Surrogate ELBO: {:.1f}".format(elbos[0]))
-
-        # Run the optimizer
-        step = dict(sgd=sgd_step, rmsprop=rmsprop_step, adam=adam_step)[optimizer]
-        state = None
-        for itr in pbar:
-            # Update the emission and variational posterior parameters
-            params, val, g, state = step(value_and_grad(_objective), params, itr, state)
-            elbos.append(-val * T)
-
-            # Update progress bar
-            pbar.set_description("Surrogate ELBO: {:.1f}".format(elbos[-1]))
-            pbar.update()
-
-        # Save the final emission and variational parameters
-        if learning:
-            self.emissions.params, variational_posterior.params = params
-        else:
-            variational_posterior.params = params
-
-        return elbos
-
-    def _fit_variational_em_with_conjugate_updates(\
-            self, variational_posterior, datas, inputs, masks, tags,
-            learning=True, alpha=.75, optimizer="adam", num_iters=100, **kwargs):
-        """
-        In the special case where the dynamics and observations are both linear
-        Gaussian, we can perform mean field coordinate ascent in a posterior
-        approximation of the form,
-
-            p(x, z | y) \approx q(x) q(z)
-
-        where q(x) is a linear Gaussian dynamical system and q(z) is a hidden
-        Markov model.
-        """
-        raise NotImplementedError
-
-    @ensure_variational_args_are_lists
-    def fit(self, variational_posterior, datas,
-            inputs=None, masks=None, tags=None, method="svi",
-            initialize=True, **kwargs):
-
-        """
-        Fitting methods for an arbitrary switching LDS:
-
-        1. Black box variational inference (bbvi/svi): stochastic gradient ascent
-           on the evidence lower bound, collapsing out the discrete states and
-           maintaining a variational posterior over the continuous states only.
-
-           Pros: simple and broadly applicable.  easy to implement.
-           Cons: doesn't leverage model structure.  slow to converge.
-
-        2. Variational expectation maximization (vem): variational posterior
-           on the continuous states q(x) and a discrete Markov chain
-           posterior on the discrete states q(z). We use samples of q(x)
-           to approximate the log transition matrix (pairwise potentials)
-           and the log transition bias (unary potentials) for q(z).  From
-           these we can derive the necessary expectations wrt q(z) for
-           updating the model parameters theta.
-
-        In the future, we could also consider some other possibilities, like:
-
-        3. Particle EM: run a (Rao-Blackwellized) particle filter targeting
-           the posterior distribution of the continuous latent states and
-           use its weighted trajectories to get the discrete states and perform
-           a Monte Carlo M-step.
-
-        4. Structured mean field: Maintain variational factors q(z) and q(x).
-           Update them using block mean field coordinate ascent, if we have a
-           Gaussian emission model and linear Gaussian dynamics, or using an
-           approximate update (e.g. a Laplace approximation) if we have a
-           nonconjugate model.
-
-        5. Gibbs sampling: As above, if we have a conjugate emission and dynamics
-           model we can do block Gibbs sampling of the discrete and continuous
-           states.
-        """
-
-        # Specify fitting methods
-        _fitting_methods = dict(svi=self._fit_svi,
-                                bbvi=self._fit_svi,
-                                vem=self._fit_variational_em)
-
-        # Deprecate "svi" as a method
-        warnings.warn("SLDS fitting method 'svi' will be renamed 'bbvi' in future releases.",
-                      category=DeprecationWarning)
-
-        if method not in _fitting_methods:
-            raise Exception("Invalid method: {}. Options are {}".\
-                            format(method, _fitting_methods.keys()))
-
-        if initialize:
-            self.initialize(datas, inputs, masks, tags)
-
-        return _fitting_methods[method](variational_posterior, datas, inputs, masks, tags,
-            learning=True, **kwargs)
-
-    @ensure_variational_args_are_lists
-    def approximate_posterior(self, variational_posterior, datas, inputs=None, masks=None, tags=None,
-                              method="svi", **kwargs):
-        # Specify fitting methods
-        _fitting_methods = dict(svi=self._fit_svi,
-                                vem=self._fit_variational_em)
-
-        if method not in _fitting_methods:
-            raise Exception("Invalid method: {}. Options are {}".\
-                            format(method, _fitting_methods.keys()))
-
-        return _fitting_methods[method](variational_posterior, datas, inputs, masks, tags,
-            learning=False, **kwargs)
-
-
-class BaseLDS(BaseSwitchingLDS):
-    """
-    Switching linear dynamical system fit with
-    stochastic variational inference on the marginal model,
-    integrating out the discrete states.
-    """
-    def __init__(self, N, D, M, dynamics, emissions):
-        from ssm.init_state_distns import InitialStateDistribution
-        from ssm.transitions import StationaryTransitions
-        init_state_distn = InitialStateDistribution(1, D, M)
-        transitions = StationaryTransitions(1, D, M)
-        super(BaseLDS, self).__init__(N, 1, D, M, init_state_distn, transitions, dynamics, emissions)
-
-    @ensure_slds_args_not_none
-    def expected_states(self, variational_mean, data, input=None, mask=None, tag=None):
-        return np.ones((variational_mean.shape[0], 1)), \
-               np.ones((variational_mean.shape[0], 1, 1)), \
-               0
-
-    @ensure_slds_args_not_none
-    def most_likely_states(self, variational_mean, data, input=None, mask=None, tag=None):
-        raise NotImplementedError
-
-    def log_prior(self):
-        return self.dynamics.log_prior() + self.emissions.log_prior()
-
-    @ensure_args_are_lists
-    def log_probability(self, datas, inputs=None, masks=None, tags=None):
-        warnings.warn("Log probability of LDS is not yet implemented.")
-        return np.nan
-
-    @ensure_variational_args_are_lists
-    def elbo(self, variational_posterior, datas, inputs=None, masks=None, tags=None, n_samples=1):
-        """
-        Lower bound on the marginal likelihood p(y | theta)
-        using variational posterior q(x; phi) where phi = variational_params
-        """
-        elbo = 0
-        for sample in range(n_samples):
-            # Sample x from the variational posterior
-            xs = variational_posterior.sample()
-
-            # log p(theta)
-            elbo += self.log_prior()
-
-            # Compute log p(y, x | theta)
-            for x, data, input, mask, tag in zip(xs, datas, inputs, masks, tags):
-                x_mask = np.ones_like(x, dtype=bool)
-                elbo += np.sum(self.dynamics.log_likelihoods(x, input, x_mask, tag))
-                elbo += np.sum(self.emissions.log_likelihoods(data, input, mask, tag, x))
-
-            # -log q(x)
-            elbo -= variational_posterior.log_density(xs)
-            assert np.isfinite(elbo)
-
-        return elbo / n_samples
