+++ tests/test_basics.py
@@ -1,7 +1,7 @@
 import autograd.numpy as np
 import autograd.numpy.random as npr
 
-from ssm.models import HMM
+import ssm
 
 
 def test_sample(T=10, K=4, D=3, M=2):
@@ -40,27 +40,27 @@ def test_sample(T=10, K=4, D=3, M=2):
     # Sample basic (no prefix, inputs, etc.)
     for transitions in transition_names:
         for observations in observation_names:
-            hmm = HMM(K, D, M=0, transitions=transitions, observations=observations)
+            hmm = ssm.HMM(K, D, M=0, transitions=transitions, observations=observations)
             zsmpl, xsmpl = hmm.sample(T)
 
     # Sample with prefix
     for transitions in transition_names:
         for observations in observation_names:
-            hmm = HMM(K, D, M=0, transitions=transitions, observations=observations)
+            hmm = ssm.HMM(K, D, M=0, transitions=transitions, observations=observations)
             zpre, xpre = hmm.sample(3)
             zsmpl, xsmpl = hmm.sample(T, prefix=(zpre, xpre))
 
     # Sample with inputs
     for transitions in transition_names:
         for observations in observation_names:
-            hmm = HMM(K, D, M=M, transitions=transitions, observations=observations)
+            hmm = ssm.HMM(K, D, M=M, transitions=transitions, observations=observations)
             zpre, xpre = hmm.sample(3, input=npr.randn(3, M))
             zsmpl, xsmpl = hmm.sample(T, prefix=(zpre, xpre), input=npr.randn(T, M))
 
     # Sample without noise
     for transitions in transition_names:
         for observations in observation_names:
-            hmm = HMM(K, D, M=M, transitions=transitions, observations=observations)
+            hmm = ssm.HMM(K, D, M=M, transitions=transitions, observations=observations)
             zpre, xpre = hmm.sample(3, input=npr.randn(3, M))
             zsmpl, xsmpl = hmm.sample(T, prefix=(zpre, xpre), input=npr.randn(T, M), with_noise=False)
 
@@ -90,7 +90,7 @@ def test_hmm_likelihood(T=1000, K=5, D=2):
     true_lkhd = oldhmm.log_likelihood(y)
 
     # Make an HMM with these parameters
-    hmm = HMM(K, D, observations="diagonal_gaussian")
+    hmm = ssm.HMM(K, D, observations="diagonal_gaussian")
     hmm.transitions.log_Ps = np.log(A)
     hmm.observations.mus = C
     hmm.observations.sigmasq = sigma * np.ones((K, D))
@@ -132,7 +132,7 @@ def test_expectations(T=1000, K=20, D=2):
     true_E_trans = states.expected_transcounts
 
     # Make an HMM with these parameters
-    hmm = HMM(K, D, observations="diagonal_gaussian")
+    hmm = ssm.HMM(K, D, observations="diagonal_gaussian")
     hmm.transitions.log_Ps = np.log(A)
     hmm.observations.mus = C
     hmm.observations.sigmasq = sigma * np.ones((K, D))
@@ -174,7 +174,7 @@ def test_viterbi(T=1000, K=20, D=2):
     z_star = states.stateseq
 
     # Make an HMM with these parameters
-    hmm = HMM(K, D, observations="diagonal_gaussian")
+    hmm = ssm.HMM(K, D, observations="diagonal_gaussian")
     hmm.transitions.log_Ps = np.log(A)
     hmm.observations.mus = C
     hmm.observations.sigmasq = sigma * np.ones((K, D))
